app/batch/requirements.txt
tensorflow==2.16.1
pillow
numpy
azure-identity
azure-storage-blob
requests
